+++ eliot/_action.py
@@ -88,7 +88,7 @@ class Action(object):
 
     @ivar _successFields: Fields to be included in successful finish message.
 
-    @ivar _failureFields: Fields to be included in a failed finish message.
+    @ivar _finished: L{True} if the L{Action} has finished, otherwise L{False}.
     """
     def __init__(self, logger, task_uuid, task_level, action_type,
                  serializers=None):
@@ -117,13 +117,13 @@ class Action(object):
         self._numberOfChildren = 0
         self._numberOfMessages = iter(count())
         self._successFields = {}
-        self._failureFields = {}
         self._logger = logger
         self._identification = {"task_uuid": task_uuid,
                                 "task_level": task_level,
                                 "action_type": action_type,
                                 }
         self._serializers = serializers
+        self._finished = False
 
 
     def _incrementMessageCounter(self):
@@ -154,7 +154,7 @@ class Action(object):
         Message(fields, serializer).write(self._logger, self)
 
 
-    def _finish(self, exception=None):
+    def finish(self, exception=None):
         """
         Log the finish message.
 
@@ -165,10 +165,13 @@ class Action(object):
         block or L{Action.finishAfter}.
 
         @param exception: C{None}, in which case the fields added with
-            L{Action.addSuccessFields} are used. Or an L{Exception}, in which case
-            an C{"exception"} field is added with the given L{Exception} and the
-            fields added with L{Action.addFailureFields} are used.
+            L{Action.addSuccessFields} are used. Or an L{Exception}, in
+            which case an C{"exception"} field is added with the given
+            L{Exception} type and C{"reason"} with its contents.
         """
+        if self._finished:
+            return
+        self._finished = True
         serializer = None
         if exception is None:
             fields = self._successFields
@@ -176,7 +179,7 @@ class Action(object):
             if self._serializers is not None:
                 serializer = self._serializers.success
         else:
-            fields = self._failureFields
+            fields = {}
             fields["exception"] = "%s.%s" % (exception.__class__.__module__,
                                              exception.__class__.__name__)
             fields["reason"] = safeunicode(exception)
@@ -258,7 +261,7 @@ class Action(object):
                 exception = result.value
             else:
                 exception = None
-            self._finish(exception)
+            self.finish(exception)
             return result
         deferred.addBoth(done)
 
@@ -271,15 +274,6 @@ class Action(object):
         @param fields: Additional fields to add to the result message.
         """
         self._successFields.update(fields)
-
-
-    def addFailureFields(self, **fields):
-        """
-        Add fields to be included in the result message if the action fails.
-
-        @param fields: Additional fields to add to the result message.
-        """
-        self._failureFields.update(fields)
 
 
     # Python context manager implementation:
@@ -296,7 +290,7 @@ class Action(object):
         Pop this action off the execution context, log finish message.
         """
         _context.pop()
-        self._finish(exception)
+        self.finish(exception)