+++ eliot/_validation.py
@@ -301,7 +301,7 @@ class ActionType(object):
         RESULT = Field("result", [str], u"The result of lookups.")
         LOG_DOSOMETHING = ActionType(
             "yourapp:subsystem:youraction",
-            [KEY], [RESULT], [],
+            [KEY], [RESULT],
             u"Do something with a key, resulting in a value.")
 
         # Actual code, with logging added:
@@ -335,8 +335,7 @@ class ActionType(object):
     _startTask = staticmethod(startTask)
 
 
-    def __init__(self, action_type, startFields, successFields, failureFields,
-                 description):
+    def __init__(self, action_type, startFields, successFields, description):
         self.action_type = action_type
         self.description = description
 
@@ -348,7 +347,7 @@ class ActionType(object):
             actionTypeField, makeActionStatusField("started")]
         successFields = successFields + [
             actionTypeField, makeActionStatusField("succeeded")]
-        failureFields = failureFields + [
+        failureFields = [
             actionTypeField, makeActionStatusField("failed"), REASON,
             EXCEPTION]