+++ eliot/tests/test_action.py
@@ -172,7 +172,7 @@ class ActionTests(TestCase):
         """
         serializers = ActionType("sys:thename",
                                  [Field("key", lambda x: x, "")],
-                                 [], [], "")._serializers
+                                 [], "")._serializers
         class Logger(list):
             def write(self, msg, serializer):
                 self.append(serializer)
@@ -336,12 +336,12 @@ class ActionTests(TestCase):
 
     def test_finish(self):
         """
-        L{Action._finish} with no exception logs an
-        C{action_status="succeeded"} message.
+        L{Action.finish} with no exception logs an C{action_status="succeeded"}
+        message.
         """
         logger = MemoryLogger()
         action = Action(logger, "unique", "/", "sys:thename")
-        action._finish()
+        action.finish()
         assertContainsFields(self, logger.messages[0],
                              {"task_uuid": "unique",
                               "task_level": "/",
@@ -351,28 +351,11 @@ class ActionTests(TestCase):
 
     def test_successfulFinishSerializer(self):
         """
-        L{Action._finish} with no exception passes the success
+        L{Action.finish} with no exception passes the success
         L{eliot._validation._MessageSerializer} to the message it creates.
         """
         serializers = ActionType("sys:thename",
                                  [],
-                                 [Field("key", lambda x: x, "")],
-                                 [], "")._serializers
-        class Logger(list):
-            def write(self, msg, serializer):
-                self.append(serializer)
-        logger = Logger()
-        action = Action(logger, "unique", "/", "sys:thename", serializers)
-        action._finish()
-        self.assertIs(logger[0], serializers.success)
-
-
-    def test_failureFinishSerializer(self):
-        """
-        L{Action._finish} with an exception passes the failure
-        L{eliot._validation._MessageSerializer} to the message it creates.
-        """
-        serializers = ActionType("sys:thename", [], [],
                                  [Field("key", lambda x: x, "")],
                                  "")._serializers
         class Logger(list):
@@ -380,25 +363,42 @@ class ActionTests(TestCase):
                 self.append(serializer)
         logger = Logger()
         action = Action(logger, "unique", "/", "sys:thename", serializers)
-        action._finish(Exception())
+        action.finish()
+        self.assertIs(logger[0], serializers.success)
+
+
+    def test_failureFinishSerializer(self):
+        """
+        L{Action.finish} with an exception passes the failure
+        L{eliot._validation._MessageSerializer} to the message it creates.
+        """
+        serializers = ActionType("sys:thename", [],
+                                 [Field("key", lambda x: x, "")],
+                                 "")._serializers
+        class Logger(list):
+            def write(self, msg, serializer):
+                self.append(serializer)
+        logger = Logger()
+        action = Action(logger, "unique", "/", "sys:thename", serializers)
+        action.finish(Exception())
         self.assertIs(logger[0], serializers.failure)
 
 
     def test_startFieldsNotInFinish(self):
         """
-        L{Action._finish} logs a message without the fields from
+        L{Action.finish} logs a message without the fields from
         L{Action._start}.
         """
         logger = MemoryLogger()
         action = Action(logger, "unique", "/", "sys:thename")
         action._start({"key": "value"})
-        action._finish()
+        action.finish()
         self.assertNotIn("key", logger.messages[1])
 
 
     def test_finishWithBadException(self):
         """
-        L{Action._finish} still logs a message if the given exception raises
+        L{Action.finish} still logs a message if the given exception raises
         another exception when called with C{unicode()}.
         """
         logger = MemoryLogger()
@@ -406,7 +406,7 @@ class ActionTests(TestCase):
         class BadException(Exception):
             def __str__(self):
                 raise TypeError()
-        action._finish(BadException())
+        action.finish(BadException())
         self.assertEqual(logger.messages[0]["reason"],
                          "eliot: unknown, unicode() raised exception")
 
@@ -482,56 +482,6 @@ class ActionTests(TestCase):
                              {"x": 1, "y": 2, "z": 3})
 
 
-    def test_addSuccessFieldsIgnoresFailure(self):
-        """
-        On a successful finish, L{Action.__exit__} ignores fields from
-        L{Action.addFailureFields}.
-        """
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", "/1/", "sys:me")
-        with action as act:
-            act.addFailureFields(z=3)
-        self.assertNotIn("z", logger.messages[0])
-
-
-    def test_addFailureFields(self):
-        """
-        On an failed finish, L{Action.__exit__} adds fields from
-        L{Action.addFailureFields} to the result message.
-        """
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", "/1/", "sys:me")
-        try:
-            with action as act:
-                act.addFailureFields(x=1, y=2)
-                act.addFailureFields(z=3)
-                raise RuntimeError()
-        except RuntimeError:
-            pass
-        else:
-            self.fail("No exception")
-        assertContainsFields(self, logger.messages[0],
-                             {"x": 1, "y": 2, "z": 3})
-
-
-    def test_addFailureFieldsIgnoresSuccess(self):
-        """
-        On a successful finish, L{Action.__exit__} ignores fields from
-        L{Action.addFailureFields}.
-        """
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", "/1/", "sys:me")
-        try:
-            with action as act:
-                act.addSuccessFields(z=3)
-                raise RuntimeError()
-        except RuntimeError:
-            pass
-        else:
-            self.fail("No exception")
-        self.assertNotIn("z", logger.messages[0])
-
-
     def test_incrementMessageCounter(self):
         """
         Each call to L{Action._incrementMessageCounter} increments a counter.
@@ -539,6 +489,21 @@ class ActionTests(TestCase):
         action = Action(MemoryLogger(), "uuid", "/1/", "sys:me")
         self.assertEqual([action._incrementMessageCounter() for i in range(5)],
                          list(range(5)))
+
+
+    def test_multipleFinishCalls(self):
+        """
+        If L{Action.finish} is called, subsequent calls to L{Action.finish} have
+        no effect.
+        """
+        logger = MemoryLogger()
+        action = Action(logger, "uuid", "/1/", "sys:me")
+        with action as act:
+            act.finish()
+            act.finish(Exception())
+            act.finish()
+        # Only initial finish message is logged:
+        self.assertEqual(len(logger.messages), 1)
 
 
 
@@ -600,13 +565,11 @@ class TwistedActionTests(TestCase):
     def test_finishAfterFailure(self):
         """
         When the L{Deferred} passed to L{Action.finishAfter} fires
-        with an exception, a finish message is logged with any failure bindings
-        added.
+        with an exception, a finish message is logged.
         """
         d = Deferred()
         logger = MemoryLogger()
         action = Action(logger, "uuid", "/1/", "sys:me")
-        action.addFailureFields(x=2)
         action.finishAfter(d)
         exception = RuntimeError("because")
         d.errback(exception)
@@ -615,7 +578,6 @@ class TwistedActionTests(TestCase):
                               "task_level": "/1/",
                               "action_type": "sys:me",
                               "action_status": "failed",
-                              "x": 2,
                               "reason": "because",
                               "exception":
                               "%s.RuntimeError" % (RuntimeError.__module__,)})